Feat : Order controller 1차 구현

+++ src/main/java/com/konkuk/bit/bitpay/order/domain/OrderDetail.java
@@ -20,7 +20,7 @@ public class OrderDetail {
     private Integer quantity;
 
     @Column(name = "order_detail_menu_id")
-    private Long menuId;
+    private Integer menuId;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
